Divide extensions into a separate files.
